vision-service/requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Requests
requests==2.31.0

# Image Processing
Pillow==10.2.0

# Data Validation
pydantic==2.5.3

# CORS Support (included in FastAPI but explicit)
python-multipart==0.0.6

# Optional but recommended
python-dotenv==1.0.0  # For environment variables